# lib/api-zod/src/generated/types/agreementStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Discourse Annotation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface AgreementStats {
  postsWithMultipleAnnotations: number;
  anthropomorphismAgreementPct: number;
  mindPerceptionAgreementPct: number;
  moralEvaluationAgreementPct: number;
  uncannyAgreementPct: number;
}