# lib/api-zod/src/generated/types/annotation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Discourse Annotation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Annotation {
  id: number;
  postId: number;
  coderId: number;
  /** @nullable */
  coderName?: string | null;
  /** none | mild | strong */
  anthropomorphismLevel: string;
  /** agency | experience | both | neither */
  mindPerception: string;
  /** praise | blame | concern | ambivalent | none */
  moralEvaluation: string;
  vassValues: boolean;
  vassAutonomy: boolean;
  vassSocialConnection: boolean;
  vassSelfAwareEmotions: boolean;
  /** eerie | creepy | fake-human | unsettling | none */
  uncanny: string;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}