# lib/api-zod/src/generated/types/bulkCreatePostsBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Discourse Annotation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePostBody } from "./createPostBody";

export interface BulkCreatePostsBody {
  posts: CreatePostBody[];
}