# lib/api-zod/src/generated/types/coder.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Discourse Annotation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Coder {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  /** lead | coder */
  role: string;
  annotationCount: number;
  createdAt: string;
}