# lib/api-zod/src/generated/types/createAnnotationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Discourse Annotation Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAnnotationBody {
  postId: number;
  coderId: number;
  anthropomorphismLevel: string;
  mindPerception: string;
  moralEvaluation: string;
  vassValues: boolean;
  vassAutonomy: boolean;
  vassSocialConnection: boolean;
  vassSelfAwareEmotions: boolean;
  uncanny: string;
  /** @nullable */
  notes?: string | null;
}