# lib/api-zod/src/generated/types/statsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Discourse Annotation Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { StatsSummaryAnthropomorphismBreakdown } from "./statsSummaryAnthropomorphismBreakdown";
import type { StatsSummaryMindPerceptionBreakdown } from "./statsSummaryMindPerceptionBreakdown";
import type { StatsSummaryMoralEvaluationBreakdown } from "./statsSummaryMoralEvaluationBreakdown";

export interface StatsSummary {
  totalPosts: number;
  annotatedPosts: number;
  unannotatedPosts: number;
  totalAnnotations: number;
  totalCoders: number;
  anthropomorphismBreakdown: StatsSummaryAnthropomorphismBreakdown;
  moralEvaluationBreakdown: StatsSummaryMoralEvaluationBreakdown;
  mindPerceptionBreakdown: StatsSummaryMindPerceptionBreakdown;
}